update BitMask. tests. assertions.

+++ Rei/src/Ecs/TypeId.h
@@ -6,14 +6,14 @@ namespace rei::ecs
     {
     public:
         template <typename>
-        static u32 Get()
+        static u64 Get()
         {
             static const u32 ID = Allocate();
             return ID;
         }
 
     private:
-        static u32 Allocate()
+        static u64 Allocate()
         {
             static u32 id = 0;
             return id++;
